configuration notte ave pourcentage

+++ backend/gestion/src/main/java/com/entreprise/gestion/rh/dto/EvaluationResultDTO.java
@@ -7,5 +7,5 @@ import java.util.Map;
 @Data
 public class EvaluationResultDTO {
     private CandidatDTO candidat;
-    private Map<String, Integer> scores;
+    private Map<String, Double> scores;
 }
